+++ src/estimators/bloom_filters.py
@@ -22,6 +22,8 @@ import numpy as np
 from scipy import special
 
 from wfa_cardinality_estimation_evaluation_framework.estimators import any_sketch
+from wfa_cardinality_estimation_evaluation_framework.estimators.estimator_noisers import GeometricEstimateNoiser
+from wfa_cardinality_estimation_evaluation_framework.common import noisers
 from wfa_cardinality_estimation_evaluation_framework.estimators.base import EstimatorBase
 from wfa_cardinality_estimation_evaluation_framework.estimators.base import SketchNoiserBase
 
@@ -417,6 +419,58 @@ class FirstMomentEstimator(EstimatorBase):
         union, self._weights)]
     return [FirstMomentEstimator._estimate_cardinality_any(
         union, self._weights)]
+
+
+class FirstMomentGlobalNoiseEstimator(EstimatorBase):
+  """This estimator unions the input sketches, generate geometric noise,
+     appends that many fake buckets to the sketch union and calls the
+     FirstMomentEstimator.
+  """
+
+  # TODO(kungfucraig): Figure out how to get the framework to pass
+  # a random state.
+  def __init__(self, epsilon, method, random_state=None):
+    self.epsilon = epsilon
+    assert method == FirstMomentEstimator.METHOD_EXP, "only METHOD_EXP is supported."
+    self.estimator = FirstMomentEstimator(method=method)
+    self.random_state = random_state
+
+  def __call__(self, sketch_list):
+    if not sketch_list:
+      return 0
+    union = sketch_list[0]
+    for cur_sketch in sketch_list[1:]:
+      union.assert_compatible(cur_sketch)
+      union.sketch = 1 - (1 - union.sketch) * (1 - cur_sketch.sketch)
+
+    noiser = GeometricEstimateNoiser(
+        self.epsilon, random_state=self.random_state)
+    noise = noiser(0.)
+    print(noiser(0.0))
+    print(noiser(0.0))
+    print(noiser(0.0))
+    print(noiser(0.0))
+
+    a = union.decay_rate
+    def _expected_num_bits(reach):
+      """Expected number of bits activated for cardinality."""
+      if reach <= 0:
+        return 0
+      return 1 - (- special.expi(- a * reach / (np.exp(a) - 1)) +
+                  special.expi(- a * np.exp(a) * reach / (np.exp(a) - 1))) / a
+
+    def _clip(x, lower_bound, upper_bound):
+      return max(min(x, upper_bound), lower_bound)
+
+    print("******** noise: %d" % noise)
+    x = sum(union.sketch) + noise
+    if x < 0:
+        x = 0
+    m = len(union.sketch)
+    p = _clip(x / m, 0, 1)
+    result = invert_monotonic(_expected_num_bits, epsilon=1e-7)(p) * m
+    assert result >= 0, "Negative estimate should never happen."
+    return int(result)
 
 
 class FixedProbabilityBitFlipNoiser(SketchNoiserBase):
